Add support for using current date when logging entry w/ no date
This adds support for automatically using today's date if the user does
not enter a date when using the 'log' function. I'm not totally happy
with the way I modified the "CheckUserEnteredProperDate" to essentially
ignore the fact the user didn't specify -d at the commandline, which was
necessary to have the program continue so that today's date could be
used later on. Definitely open to suggestions if there's a better way to
do that!